chat and warehouse pages
