+++ lib/src/home/widgets/explore_slider.dart
@@ -1,5 +1,5 @@
 import 'package:bookmywarehouse/constants/color/base_color.dart';
-import 'package:bookmywarehouse/widgets/propert_details.dart';
+import 'package:bookmywarehouse/src/warehouse/pages/propert_details.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
@@ -22,7 +22,7 @@ class _ExploreSliderState extends State<ExploreSlider> {
     var width = MediaQuery.of(context).size.width;
 
     return Container(
-      height: height * 0.36,
+      height: height * 0.39,
       width: width,
       margin: EdgeInsets.symmetric(
         horizontal: width * 0.02,
