+++ lib/src/home/widgets/explore_slider_two.dart
@@ -1,59 +1,67 @@
 import 'package:bookmywarehouse/constants/color/base_color.dart';
-import 'package:bookmywarehouse/widgets/propert_details.dart';
+import 'package:bookmywarehouse/src/warehouse/pages/propert_details.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
-class CustomSliderTwoHome extends StatefulWidget {
-  const CustomSliderTwoHome({
+class ExploreSliderTwo extends StatefulWidget {
+  const ExploreSliderTwo({
     super.key,
     required this.title,
+    required this.subTitle2,
   });
-  final String title;
+  final String title, subTitle2;
 
   @override
-  State<CustomSliderTwoHome> createState() => _CustomSliderTwoHomeState();
+  State<ExploreSliderTwo> createState() => _ExploreSliderTwoState();
 }
 
-class _CustomSliderTwoHomeState extends State<CustomSliderTwoHome> {
+class _ExploreSliderTwoState extends State<ExploreSliderTwo> {
   @override
   Widget build(BuildContext context) {
     var height = MediaQuery.of(context).size.height;
     var width = MediaQuery.of(context).size.width;
 
     return Container(
-      height: height * 0.42,
+      height: height * 0.36,
       width: width,
-      margin: EdgeInsets.symmetric(horizontal: width * 0.02),
+      margin: EdgeInsets.symmetric(
+        horizontal: width * 0.02,
+        vertical: width * 0.02,
+      ),
       child: Column(
         children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              SizedBox(
-                height: height * 0.08,
-                width: width * 0.5,
-                child: Text(
-                  widget.title,
-                  style: GoogleFonts.poppins(
-                    textStyle: TextStyle(
-                        fontSize: 18,
-                        letterSpacing: 0.13,
-                        color: BasicColor.deepBlack,
-                        fontWeight: FontWeight.w800),
+          Padding(
+            padding: EdgeInsets.symmetric(horizontal: width * 0.02),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                SizedBox(
+                  width: width * 0.7,
+                  height: height * 0.06,
+                  child: Text(
+                    widget.title,
+                    style: GoogleFonts.poppins(
+                      textStyle: TextStyle(
+                          fontSize: 18,
+                          letterSpacing: 0.13,
+                          color: BasicColor.deepBlack,
+                          fontWeight: FontWeight.w800),
+                    ),
                   ),
                 ),
-              ),
-              TextButton(
-                onPressed: () {
-                  print('see all clicked');
-                },
-                child: const Text('See all'),
-              )
-            ],
+                TextButton(
+                  onPressed: () {
+                    print('see all clicked');
+                  },
+                  child: const Text('See all'),
+                )
+              ],
+            ),
           ),
-          SizedBox(
+          Container(
             width: width,
             height: height * 0.28,
+            margin: EdgeInsets.only(top: 10),
             child: ListView.builder(
               scrollDirection: Axis.horizontal,
               itemCount: 10,
@@ -103,7 +111,7 @@ class _CustomSliderTwoHomeState extends State<CustomSliderTwoHome> {
                               height: height * 0.02,
                               width: width * 0.5,
                               child: Text(
-                                '345 rented props ',
+                                widget.subTitle2,
                                 style: GoogleFonts.poppins(
                                     textStyle: const TextStyle(
                                   fontSize: 14,
