+++ lib/src/home/widgets/switch_bar.dart
