+++ lib/src/rent/widgets/checkout_appbar.dart
