+++ lib/src/rent/widgets/house_details.dart
