+++ lib/src/rent/widgets/input_details.dart
