+++ lib/src/rent/widgets/price_details.dart
