+++ lib/src/warehouse/repository/warehouse_repository.dart
@@ -1,0 +1,6 @@
+import 'package:bookmywarehouse/common/failure/failure.dart';
+import 'package:fpdart/fpdart.dart';
+
+abstract class WarehouseRepository {
+  Future<Either<Failure, Map<String, dynamic>>> fetchWarehouse(String uuid);
+}
