+++ lib/src/warehouse/widgets/property_details_slider.dart
