+++ lib/src/warehouse/widgets/rent_button.dart
@@ -1,5 +1,5 @@
 import 'package:bookmywarehouse/constants/color/base_color.dart';
-import 'package:bookmywarehouse/widgets/checkout.dart';
+import 'package:bookmywarehouse/src/rent/pages/checkout.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
