+++ lib/src/warehouse/widgets/warehouse_details.dart
